chore(server): move reset token service
